requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
dvc
hydra-core
pandas
pre-commit
swig >= 4.1.1
smac >= 2.0.0
numpy
matplotlib
ConfigSpace
